move motor to pkg
