+++ pkg/motor/motor.go
