+++ pkg/motor/motor_test.go
