+++ pkg/motor/proto/common/v1/info.proto
