+++ pkg/motor/x/transmit/session.go
